+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `sorbetto.enablePackageSupport` setting to enable LSP package support via `--sorbet-packages` flag.
 - Add **Sobetto: Send Selection to sorbet.run** command (1MB cap).
 - **Sorbetto: Debug Ruby File** attempts to launch the best debugger possible (executable or VSCode debugger-type).
+- Expanded README content, especially around configuration.
 
 ## 0.3.15
 - New **Sorbetto: Update RBIs** command to manage RBIs using `tapioca`.